Bring goToRandomArticle back to Search
