+++ src/utils/goto.js
@@ -1,6 +1,5 @@
 import { route } from 'preact-router'
 import { canonicalizeTitle, getAppLanguage } from 'utils'
-import { getRandomArticleTitle } from 'api'
 
 const article = (lang, title, replace = false) => {
   if (!Array.isArray(title)) {
@@ -27,18 +26,6 @@ const privacyPolicy = () => {
   window.open('https://foundation.m.wikimedia.org/wiki/Privacy_policy')
 }
 
-const randomArticle = (closePopup) => {
-  const lang = getAppLanguage()
-  const [promise] = getRandomArticleTitle(lang)
-
-  promise.then(title => {
-    if (closePopup) {
-      closePopup()
-    }
-    article(lang, title)
-  })
-}
-
 const back = () => window.history.back()
 
-export const goto = { article, search, consent, termsOfUse, privacyPolicy, back, randomArticle, settings, tips }
+export const goto = { article, search, consent, termsOfUse, privacyPolicy, back, settings, tips }
